restructure journal files
