+++ alpaca/broker/models/__init__.py
@@ -4,3 +4,4 @@ from .requests import *
 from .documents import *
 from .funding import *
 from .trading import *
+from .journals import *
